+++ server/src/main/java/org/allaymc/server/block/component/button/BlockButtonBaseComponentImpl.java
@@ -63,7 +63,7 @@ public class BlockButtonBaseComponentImpl extends BlockBaseComponentImpl {
         var clickedBlockState = interactInfo.getClickedBlockState();
         if (!clickedBlockState.getPropertyValue(BUTTON_PRESSED_BIT)) {
             dimension.updateBlockProperty(BUTTON_PRESSED_BIT, true, pos);
-            dimension.getBlockUpdateService().scheduleBlockUpdate(pos, getActivationTime());
+            dimension.getBlockUpdateService().scheduleBlockUpdateInDelay(pos, getActivationTime());
             dimension.addLevelSoundEvent(pos.x() + 0.5f, pos.y() + 0.5f, pos.z() + 0.5f, SoundEvent.BUTTON_CLICK_ON);
         }
         return true;